feat: add MenuBarProcessGroupRow and MenuBarTunnelRow views
- Implemented MenuBarProcessGroupRow to display process groups with options to expand, kill processes, and show status indicators.
- Created MenuBarTunnelRow to represent Cloudflare tunnel states with context menu options for copying URLs and stopping tunnels.
- Refactored PortDetailView to simplify tunnel status display.
- Modularized ConnectionEditSection into smaller components: ConnectionInfoSection, PortMappingSection, and OptionsSection for better readability and maintainability.
- Introduced SearchablePickerView for selecting namespaces and services with search functionality.
- Updated PortForwardConnectionCard to utilize the new edit form structure.
